requirements.txt
beartype==0.17.2
einops==0.7.0
hydra-core==1.3.2
jaxtyping==0.2.28
librosa==0.10.1
lightning==2.2.1
matplotlib==3.8.3
mlflow==2.11.3
numpy==1.26.4
phonemizer==3.2.1
polars==0.20.13
pykakasi==2.2.1
pyopenjtalk==0.3.3
soundfile==0.12.1
tensorboard==2.16.2
torch==2.2.1
torchaudio==2.2.1
torchdiffeq==0.2.3
vocos==0.1.0
wandb==0.16.4